pyproject.toml
[project]
name = "entra-id-spotter"
version = "0.1.0"
description = "CLI tool to audit Entra ID for security misconfigurations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "msgraph-sdk>=1.0.0",
    "azure-identity>=1.15.0",
    "click>=8.0.0",
]

[project.scripts]
entra-spotter = "entra_spotter.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/entra_spotter"]
